Reorganize to differentiate old and new website
Rename the original folder for my personal website to "personal".

Created a new folder named "m.personal" and moved the files for the new website in there.

Also reorganized the layout so old files for testing are now all in the same folder. The actual final version of the website are moved to the top layer.